PC Monitor/requirements.txt
python==3.10
discord.py==2.3.2
discord
python-dotenv==1.0.0
Pillow==10.2.0
sounddevice==0.4.6
scipy==1.12.0
numpy==1.26.4
psutil==5.9.8
opencv-python==4.9.0.80
pynput==1.7.6
pyautogui==0.9.54
pyobjc-framework-Quartz==9.2; sys_platform == 'darwin'
pyobjc-framework-Cocoa==9.2; sys_platform == 'darwin' 